Avalon

+++ Avalon/Avalon.Draw/Draw.cs
@@ -200,7 +200,7 @@ public class Draw : Any
     protected virtual Form WorldForm { get; set; }
     protected virtual Form FormA { get; set; }
     protected virtual Pos PosA { get; set; }
-    protected virtual int TextCount { get; set; }
+    protected virtual long TextCount { get; set; }
     private ulong Intern { get; set; }
     private ulong InternFillPos { get; set; }
     private ulong InternArea { get; set; }
